+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## TODO
 
+* Fix request abort on view change
 * Custom names for custom views
 * Handle visibility and permanently_hidden device attributes
 * Handle rooms (http://zway-server:8083/ZAutomation/api/v1/locations) and dashboard
@@ -11,6 +12,10 @@
  - Battery low
  - Failed devices
  - T below 5C
+ 
+## Version 1.0.4+5
+
+* Added support for window blinds (SwitchMultilevel + motor)
 
 ## Version 1.0.3+4